update to async thunks and removed depricated actions
