+++ vite.config.ts
@@ -20,7 +20,6 @@ export default defineConfig(({ command, isPreview }): UserConfig => {
     root: dev ? path.resolve(__dirname, './preview') : path.resolve(__dirname, ''),
     server: {
       port: 3000,
-      origin: 'bs-local.com',
     },
     css: {
       preprocessorOptions: {
